About Us
The About us page was changed a little bit, by adding a footer which all the other pages has and also since one of the photo didn't fit into the size fixed, it was cropped and then was fitted into the size. These are the changes which were made for about us.