Member Folder
This folder contains many subfolders as well as a php file. 
Member.php: The interface for the member when he/she logins.
Change Password Folder: It has two php files. They are: changePassword.php and changeAction.php. This is used to change the password of a member.
Current Transaction Folder: This has one php file, which is named as 'CurrentDetails.php', which is used to show the books which are borrowed by the member currently and has not been returned yet.
Manage Books Folder: This has five php files and an image. The five php files are: AllBooks.php, BorrowableBooks.php, generate_image.php, ReferableBooks.php and SearchBook.php. These are used to show all the books that are available in the library and the books which are either borrowable or referable, and also to search a book with name or an book ID and also does the action to reserve the book online to the members.
Past Borrow Folder: has one php file named as 'PastBorrow.php' which shows all the borrowed details of the member which is done by him in the past.
Personal Info Folder: Which has 2 php files and an image. The php files are: 'EditAction.php' and 'PersonalInfo.php'. This allows the member to view their personal information and then the member can only edit the username or the email. The other details should be changed by verifying with the librarian.